adding a @voxel51/plugins plugin
