+++ plugins/utils/__init__.py
@@ -5,29 +5,16 @@ Utility operators.
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
-from collections import defaultdict
 import json
-
-try:
-    from importlib import metadata
-except ImportError:
-    import importlib_metadata as metadata
-
 import multiprocessing
-from packaging.requirements import Requirement
 
 import fiftyone as fo
-import fiftyone.constants as foc
 import fiftyone.core.media as fom
 import fiftyone.core.metadata as fomm
 import fiftyone.core.utils as fou
 import fiftyone.operators as foo
 import fiftyone.operators.types as types
 import fiftyone.utils.image as foui
-import fiftyone.plugins as fop
-import fiftyone.zoo as foz
-import fiftyone.zoo.datasets as fozd
-import fiftyone.zoo.models as fozm
 
 
 class CreateDataset(foo.Operator):
@@ -1077,342 +1064,6 @@ def _generate_thumbnails_inputs(ctx, inputs):
     return True
 
 
-class ManagePlugins(foo.Operator):
-    @property
-    def config(self):
-        return foo.OperatorConfig(
-            name="manage_plugins",
-            label="Manage plugins",
-            light_icon="/assets/icon-light.svg",
-            dark_icon="/assets/icon-dark.svg",
-            dynamic=True,
-        )
-
-    def resolve_input(self, ctx):
-        inputs = types.Object()
-        _manage_plugins_inputs(ctx, inputs)
-        return types.Property(inputs, view=types.View(label="Manage plugins"))
-
-    def execute(self, ctx):
-        tab = ctx.params.get("tab", None)
-        if tab == "PLUGINS":
-            _plugin_enablement(ctx)
-        elif tab == "INSTALL":
-            _install_plugin(ctx)
-
-
-def _manage_plugins_inputs(ctx, inputs):
-    tab_choices = types.TabsView()
-    tab_choices.add_choice("PLUGINS", label="Plugins")
-    tab_choices.add_choice("REQUIREMENTS", label="Requirements")
-    tab_choices.add_choice("INSTALL", label="Install")
-    default = "PLUGINS"
-
-    inputs.enum(
-        "tab",
-        tab_choices.values(),
-        default=default,
-        view=tab_choices,
-    )
-    tab = ctx.params.get("tab", default)
-
-    if tab == "PLUGINS":
-        _plugin_enablement_inputs(ctx, inputs)
-    elif tab == "REQUIREMENTS":
-        _plugin_requirements_inputs(ctx, inputs)
-    elif tab == "INSTALL":
-        _install_plugin_inputs(ctx, inputs)
-
-
-def _plugin_enablement_inputs(ctx, inputs):
-    obj = types.Object()
-    obj.str(
-        "name",
-        default="**Name**",
-        view=types.MarkdownView(read_only=True, space=3),
-    )
-    obj.str(
-        "description",
-        default="**Description**",
-        view=types.MarkdownView(read_only=True, space=7),
-    )
-    obj.str(
-        "enabled",
-        default="**Enabled**",
-        view=types.MarkdownView(read_only=True, space=2),
-    )
-    inputs.define_property("enablement_header", obj)
-
-    enabled_plugins = set(fop.list_enabled_plugins())
-
-    num_edited = 0
-    for i, plugin in enumerate(fop.list_plugins(enabled="all"), 1):
-        prop_name = f"enablement{i}"
-        actual_enabled = plugin.name in enabled_plugins
-        enabled = ctx.params.get(prop_name, {}).get("enabled", actual_enabled)
-        edited = enabled != actual_enabled
-        num_edited += int(edited)
-
-        obj = types.Object()
-        obj.str(
-            "markdown_name",
-            default=f"[{plugin.name}]({plugin.url})",
-            view=types.MarkdownView(read_only=True, space=3),
-        )
-        obj.str(
-            "description",
-            default=plugin.description,
-            view=types.MarkdownView(read_only=True, space=6.5),
-        )
-        obj.str(
-            "name",
-            default=plugin.name,
-            view=types.HiddenView(read_only=True, space=0.5),
-        )
-        obj.bool(
-            "enabled",
-            label="(edited)" if edited else "",
-            default=actual_enabled,
-            view=types.CheckboxView(space=2),  # @todo use SwitchView
-        )
-        inputs.define_property(prop_name, obj)
-
-    if num_edited > 0:
-        view = types.Notice(
-            label=(
-                f"You are about to change the enablement of {num_edited} "
-                "plugins"
-            )
-        )
-    else:
-        view = types.Notice(label="You have not made any changes")
-
-    status_prop = inputs.view("enablement_status", view)
-
-    if num_edited == 0:
-        status_prop.invalid = True
-
-
-def _plugin_enablement(ctx):
-    enabled_plugins = set(fop.list_enabled_plugins())
-
-    i = 0
-    while True:
-        i += 1
-        prop_name = f"enablement{i}"
-        obj = ctx.params.get(prop_name, None)
-        if obj is None:
-            break
-
-        name = obj["name"]
-        enabled = obj["enabled"]
-
-        actual_enabled = name in enabled_plugins
-        if enabled != actual_enabled:
-            if enabled:
-                fop.enable_plugin(name)
-            else:
-                fop.disable_plugin(name)
-
-
-def _plugin_requirements_inputs(ctx, inputs):
-    plugins = [p.name for p in fop.list_plugins(enabled="all")]
-
-    plugin_choices = types.DropdownView()
-    for name in sorted(plugins):
-        plugin_choices.add_choice(name, label=name)
-
-    inputs.str(
-        "requirements_name",
-        default=None,
-        required=True,
-        label="Plugin",
-        description="Choose a plugin whose requirements you want to check",
-        view=plugin_choices,
-    )
-
-    name = ctx.params.get("requirements_name", None)
-    if name is None:
-        return
-
-    requirements = []
-
-    plugin = fop.get_plugin(name)
-    req_str = plugin.fiftyone_requirement
-    if req_str is not None:
-        requirements.append(_check_fiftyone_requirement(req_str))
-
-    req_strs = fop.load_plugin_requirements(name)
-    if req_strs is not None:
-        for req_str in req_strs:
-            requirements.append(_check_package_requirement(req_str))
-
-    num_requirements = len(requirements)
-    if num_requirements == 0:
-        inputs.view(
-            "requirements_status",
-            types.Notice(label="This plugin has no package requirements"),
-        )
-        return
-
-    obj = types.Object()
-    obj.str(
-        "requirements_requirement",
-        default="**Requirement**",
-        view=types.MarkdownView(read_only=True, space=5),
-    )
-    obj.str(
-        "requirements_version",
-        default="**Installed version**",
-        view=types.MarkdownView(read_only=True, space=5),
-    )
-    obj.str(
-        "requirements_satisfied",
-        default="**Satisfied**",
-        view=types.MarkdownView(read_only=True, space=2),
-    )
-    inputs.define_property("requirements_header", obj)
-
-    num_satisfied = 0
-    for i, (req, version, success) in enumerate(requirements, 1):
-        prop_name = f"requirements{i}"
-        num_satisfied += int(success)
-
-        obj = types.Object()
-        obj.str(
-            "requirement",
-            default=str(req),
-            view=types.MarkdownView(read_only=True, space=5),
-        )
-        obj.str(
-            "version",
-            default=version or "-",
-            view=types.MarkdownView(read_only=True, space=5),
-        )
-        obj.bool(
-            "satisfied",
-            default=success,
-            view=types.CheckboxView(read_only=True, space=2),
-        )
-        inputs.define_property(prop_name, obj)
-
-    if num_satisfied == num_requirements:
-        view = types.Notice(label="All package requirements are satisfied")
-    else:
-        view = types.Warning(
-            label=(
-                f"Only {num_satisfied}/{num_requirements} package "
-                "requirements are satisfied"
-            )
-        )
-
-    status_prop = inputs.view("requirements_status", view)
-    status_prop.invalid = True
-
-
-def _check_fiftyone_requirement(req_str):
-    req = Requirement(req_str)
-    version = foc.VERSION
-    success = not req.specifier or req.specifier.contains(version)
-
-    return req, version, success
-
-
-def _check_package_requirement(req_str):
-    req = Requirement(req_str)
-
-    try:
-        version = metadata.version(req.name)
-    except metadata.PackageNotFoundError:
-        version = None
-
-    success = (version is not None) and (
-        not req.specifier or req.specifier.contains(version)
-    )
-
-    return req, version, success
-
-
-def _install_plugin_inputs(ctx, inputs):
-    instructions = """
-The location to download the plugin(s) from, which can be:
-
--   A GitHub repo URL like `https://github.com/<user>/<repo>`
--   A GitHub ref like
-    `https://github.com/<user>/<repo>/tree/<branch>` or
-    `https://github.com/<user>/<repo>/commit/<commit>`
--   A GitHub ref string like `<user>/<repo>[/<ref>]`
--   A publicly accessible URL of an archive (eg zip or tar) file
-    """
-
-    """
-    inputs.str(
-        "instructions",
-        default=instructions.strip(),
-        view=types.MarkdownView(read_only=True),
-    )
-    """
-
-    inputs.str(
-        "url_or_gh_repo",
-        required=True,
-        label="URL",
-        description="The GitHub repository to download the plugin(s) from",
-    )
-
-    url_or_gh_repo = ctx.params.get("url_or_gh_repo", None)
-    if not url_or_gh_repo:
-        return
-
-    inputs.list(
-        "plugin_names",
-        types.String(),
-        default=None,
-        required=False,
-        label="Plugin names",
-        description=(
-            "An optional list of plugin names to install. By default, all "
-            "found plugins are installed"
-        ),
-    )
-
-    inputs.int(
-        "max_depth",
-        default=3,
-        label="Max depth",
-        description=(
-            "The maximum depth within the downloaded archive to search for "
-            "plugins"
-        ),
-        view=types.View(space=6),
-    )
-
-    inputs.bool(
-        "overwrite",
-        default=False,
-        label="Overwrite",
-        description=(
-            "Whether to overwrite an existing plugin with the same name if it "
-            "already exists"
-        ),
-        view=types.CheckboxView(space=6),
-    )
-
-
-def _install_plugin(ctx):
-    url_or_gh_repo = ctx.params["url_or_gh_repo"]
-    plugin_names = ctx.params.get("plugin_names", None)
-    max_depth = ctx.params.get("max_depth", 3)
-    overwrite = ctx.params.get("overwrite", False)
-
-    fop.download_plugin(
-        url_or_gh_repo,
-        plugin_names=plugin_names,
-        max_depth=max_depth,
-        overwrite=overwrite,
-    )
-
-
 def _get_fields_with_type(view, type):
     if issubclass(type, fo.Field):
         return view.get_field_schema(ftype=type).keys()
@@ -1493,4 +1144,3 @@ def register(p):
     p.register(DeleteDataset)
     p.register(ComputeMetadata)
     p.register(GenerateThumbnails)
-    p.register(ManagePlugins)
